+++ CHANGELOG.md
@@ -29,6 +29,12 @@ You can check your current version with the following command:
     ```
 
 For more information, see [UP42 Python package description](https://pypi.org/project/up42-py/).
+## 2.0.0a5
+
+**Aug 7, 2024**
+- Moved `ProductGlossary` and its dependencies to `glossary.py`.
+- Published `CollectionType` in the global namespace.
+
 ## 2.0.0a4
 
 **Aug 6, 2024**